Extracted storage extension
