+++ tests/IntegrationTestCase.php
@@ -7,6 +7,7 @@ use PhpBench\Extension\CoreExtension;
 use PhpBench\Extension\ExpressionExtension;
 use PhpBench\Extension\ReportExtension;
 use PhpBench\Extension\RunnerExtension;
+use PhpBench\Extension\StorageExtension;
 use PhpBench\Tests\Util\Workspace;
 
 class IntegrationTestCase extends TestCase
@@ -26,6 +27,7 @@ class IntegrationTestCase extends TestCase
             ExpressionExtension::class,
             CoreExtension::class,
             ReportExtension::class,
+            StorageExtension::class,
             RunnerExtension::class,
         ], array_merge([
             ExpressionExtension::PARAM_SYNTAX_HIGHLIGHTING => false,
